Split token-auth into per-step modules
Break the 570-LOC src/api/token-auth.ts into a folder of focused units:
- token-auth/constants.ts — endpoints, basic auth, PKCE/state sizes,
  redirect window. Anything that is fixed by the MELCloud backend +
  the melcloudhome-mobile client registration.
- token-auth/transport.ts — fetch helpers (`fetchPostForm`,
  `fetchRaw`), `OidcResponse` shape, header reader, cookie-jar
  hand-off, and the cookie-aware `authRequest`.
- token-auth/html-parsing.ts — pure HTML scraping
  (`extractFormAction`, `extractHiddenFields`, `extractPageRedirect`).
- token-auth/redirect.ts — `authFollowRedirects`, redirect-target
  resolution, URL resolution.
- token-auth/flow.ts — the OIDC flow primitives: `generatePKCE`,
  `par`, `submitCredentials`, `extractAuthorizationCode`,
  `tokenRequest`, plus the `TokenResponse` shape.
- token-auth/index.ts — public surface unchanged: `performTokenAuth`
  and `refreshAccessToken` orchestrate the steps above.

Pure mechanical refactor — runtime behaviour, error shapes, and the
import path consumed by `src/api/home.ts` remain identical (the
folder's `index.ts` resolves `./token-auth/index.ts`).
